% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{airportMatcher}
\alias{airportMatcher}
\title{airportMatcher}
\usage{
airportMatcher(returnDF, dataSet = "openFlights")
}
\arguments{
\item{returnDF}{output of coordinateGetter}

\item{dataSet}{airport database do you want to use: "openFlights" or "the other one", default = "openFlights}
}
\value{
a dataframe containing the address, the coordinates of the address, the IATA code of the nearest airport, the distance (in km) to the nearest airport, and the coordinates of the nearest airport
}
\description{
finds the nearest airport to a set of coordinates
}
\details{
It's a bit tough to know what airports to use, it seems all corpuses are either too detailed or not enough. For now I'm using openflights, might parametrize this later to allow for other airport lists to be selected
}
\examples{
\dontrun{
if(interactive()){
 #'  library(purrr)
 addresses <- c("9th Street, Troy, NY, USA", "The Mall, Solan, Himachal Pradesh, India")
 returnDF <- purrr::map_dfr(addresses, ~coordinateGetter(.))
 airportMatcher(returnDF)
 }
}
}
\seealso{
\code{\link[geosphere]{distm}}
}
