% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{chartMaker}
\alias{chartMaker}
\title{chartMaker}
\usage{
chartMaker(
  lat_destiantion,
  lon_destination,
  lat_origin,
  lon_origin,
  color = "f2f2f2",
  lwd = 0.8
)
}
\arguments{
\item{lat_destiantion}{latitude of destination}

\item{lon_destination}{longitude of destination}

\item{lat_origin}{latitude of origin}

\item{lon_origin}{longitude of origin}

\item{color}{what color would you like the lines to
be (in hex code, like #efefef), Default: "f2f2f2"}

\item{lwd}{desired line thickness, Default:0.8}
}
\value{
returns a map with great arcs drawn to it's destination
}
\description{
Creates a network great arcs map showing all the trips people have taken to get to the destination
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
