% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{coordinateGetter}
\alias{coordinateGetter}
\title{coordinateGetter}
\usage{
coordinateGetter(address, key)
}
\arguments{
\item{address}{physical address. Works best in this format: 13 Elm Street, Amityville, State, Country. It's most important to get the two last points correct, so either City, Country, or State, Country (since that will give us decent granularity to find the nearest airport)}

\item{key}{To access the openstreetmap API you need a valid API key. You can get it for free at https://developer.mapquest.com}
}
\value{
returns a data frame of the address, its latitude and its longitude
}
\description{
takes an address and returns the coordinates of that point
}
\details{
since we are using the free api with geocode, the matching is approximate, but good enough for our purposes.
}
\examples{
\dontrun{
if(interactive()){
 library(purrr);library(magrittr)
 addresses <- c("9th Street, Troy, NY, USA", "The Mall, Solan, Himachal Pradesh, India")
 purrr::map_dfr(addresses, ~coordinateGetter(.))

 }
}
}
\seealso{
\code{\link[ggmap]{geocode}}
}
