% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{footprintGetter}
\alias{footprintGetter}
\title{footprintGetter}
\usage{
footprintGetter(from, to, connections = NULL, class = "economy", API)
}
\arguments{
\item{from}{Origin airport code}

\item{to}{Destination airport code}

\item{connections}{are there any connections? provide as comma seperated list if more than one , Default: NULL}

\item{class}{what class the flight is in. Can be 'economy', 'premium_economy',
'business' and 'first'., Default: 'economy'}

\item{API}{API Key. Contact goclimateneutral in order to obtain your own at https://www.goclimateneutral.org/contact}
}
\value{
rturns a tibble with the origin and destination airport codes as well
as the tons of CO2 of that flight and the cost to offset, currently in SEK.
}
\description{
This function queries the goclimateneutral API and gets the values for one specific trip
}
\details{
All flights are assumed to be 'to' and 'from' only, no connections (even though
this is an underestimation). It would be easy to
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
