% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invite_locations_list}
\alias{invite_locations_list}
\title{A toy list of locations}
\format{
A data frame with 50 rows and 1 variables:
\describe{
  \item{From}{Geographic location from which participants are coming}
}
}
\usage{
invite_locations_list
}
\description{
A randomly generated list of locations. This dataset is a bit dirty in that it includes
both city/country, only country, and state/country information.
}
\keyword{datasets}
