% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_airports}
\alias{world_airports}
\title{A list of airport codes from Data-hub}
\format{
A data frame with 57206 rows and 20 variables:
\describe{
  \item{ident}{numeric unique identifier?}
  \item{ident}{unique identifier?}
  \item{type}{what kind of airport is it}
  \item{name}{airport name}
  \item{latitude_deg}{latitude}
  \item{longitude_deg}{longitude}
  \item{elevation_ft}{elevation in ft}
  \item{continent}{continent}
  \item{iso_country}{3-digit country code}
  \item{iso_region}{combination of country code and subnational code}
  \item{municipality}{municipality name, mostly city name}
  \item{scheduled_service}{unknown}
  \item{gps_code}{unknown}
  \item{iata_code}{unknown}
  \item{local_code}{unknown}
  \item{home_link}{URL for airport}
  \item{wikipedia_link}{URL for wikipedia page}
  \item{keywords}{aliases for this airport}
  \item{score}{unknown}
  \item{last_updated}{date of last update for each airport}
}
}
\source{
\url{https://ourairports.com/data/}
}
\usage{
world_airports
}
\description{
This data comes from https://ourairports.com/ downloaded on 2020-07-31.
As of this release it consists off 57206 airport codes
}
\keyword{datasets}
